refactor(arch): split kernel_function into separate module
