+++ src/arch/aarch64/kernel/mod.rs
@@ -1,5 +1,6 @@
 pub mod core_local;
 pub mod interrupts;
+pub mod kernel_stack;
 #[cfg(all(
 	not(feature = "pci"),
 	any(
@@ -15,7 +16,6 @@ pub mod scheduler;
 pub mod serial;
 #[cfg(target_os = "none")]
 mod start;
-pub mod switch;
 pub mod systemtime;
 
 use alloc::alloc::{Layout, alloc};
