+++ src/arch/x86_64/kernel/kernel_stack.rs
@@ -1,0 +1,105 @@
+use core::arch::asm;
+use core::{mem, ptr};
+
+use crate::core_local::CoreLocal;
+
+macro_rules! kernel_function_impl {
+	($kernel_function:ident($($arg:ident: $A:ident),*) { $($operands:tt)* }) => {
+		/// Executes `f` on the kernel stack.
+		#[allow(dead_code)]
+		pub unsafe fn $kernel_function<R, $($A),*>(f: unsafe extern "C" fn($($A),*) -> R, $($arg: $A),*) -> R {
+			unsafe {
+				assert!(mem::size_of::<R>() <= mem::size_of::<usize>());
+
+				$(
+					assert!(mem::size_of::<$A>() <= mem::size_of::<usize>());
+					let $arg = {
+						let mut reg = 0usize;
+						// SAFETY: $A is smaller than usize and directly fits in a register
+						// Since f takes $A as argument via C calling convention, any upper bytes do not matter.
+						ptr::write(ptr::from_mut(&mut reg).cast(), $arg);
+						reg
+					};
+				)*
+
+				let ret: u64;
+				asm!(
+					// Save user stack pointer and switch to kernel stack
+					"cli",
+					"mov r12, rsp",
+					"mov rsp, {kernel_stack_ptr}",
+					"sti",
+
+					// To make sure, Rust manages the stack in `f` correctly,
+					// we keep all arguments and return values in registers
+					// until we switch the stack back. Thus follows the sizing
+					// requirements for arguments and return types.
+					"call {f}",
+
+					// Switch back to user stack
+					"cli",
+					"mov rsp, r12",
+					"sti",
+
+					f = in(reg) f,
+					kernel_stack_ptr = in(reg) CoreLocal::get().kernel_stack.get(),
+
+					$($operands)*
+
+					// user_stack_ptr saved in r12
+					out("r12") _,
+
+					// Return argument in rax
+					out("rax") ret,
+
+					clobber_abi("C"),
+				);
+
+				// SAFETY: R is smaller than usize and directly fits in rax
+				// Since f returns R, we can safely convert ret to R
+				mem::transmute_copy(&ret)
+			}
+		}
+	};
+}
+
+kernel_function_impl!(kernel_function0() {});
+
+kernel_function_impl!(kernel_function1(arg1: A1) {
+	in("rdi") arg1,
+});
+
+kernel_function_impl!(kernel_function2(arg1: A1, arg2: A2) {
+	in("rdi") arg1,
+	in("rsi") arg2,
+});
+
+kernel_function_impl!(kernel_function3(arg1: A1, arg2: A2, arg3: A3) {
+	in("rdi") arg1,
+	in("rsi") arg2,
+	in("rdx") arg3,
+});
+
+kernel_function_impl!(kernel_function4(arg1: A1, arg2: A2, arg3: A3, arg4: A4) {
+	in("rdi") arg1,
+	in("rsi") arg2,
+	in("rdx") arg3,
+	in("rcx") arg4,
+});
+
+kernel_function_impl!(kernel_function5(arg1: A1, arg2: A2, arg3: A3, arg4: A4, arg5: A5) {
+	in("rdi") arg1,
+	in("rsi") arg2,
+	in("rdx") arg3,
+	in("rcx") arg4,
+	in("r8") arg5,
+});
+
+kernel_function_impl!(kernel_function6(arg1: A1, arg2: A2, arg3: A3, arg4: A4, arg5: A5, arg6: A6) {
+	in("rdi") arg1,
+	in("rsi") arg2,
+	in("rdx") arg3,
+	in("rcx") arg4,
+	in("r8") arg5,
+	in("r9") arg6,
+});
